requirements.txt
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
defusedxml==0.5.0
Django==1.11.9
django-allauth==0.35.0
django-braces==1.12.0
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-redis==4.8.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
ez==2.8.7
ez-setup==0.9
flake8==3.5.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
oauthlib==2.0.6
pandocfilters==1.4.2
parso==0.1.1
pew==1.1.2
pexpect==4.3.1
pickleshare==0.7.4
pipenv==9.0.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.8.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
pyzmq==16.0.4
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
Send2Trash==1.4.2
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
